Map test files 1 to 1 for each module
