+++ tests/script/test_fingerprints.py
